Study of null edges not sew cases
virtual sector detection isn't needed anymore at least for the original test cases because of 'i_join_pendant_loose_ends_by_modifying_topology'.

Of course, there're still cases in which the sewing fails, but don't know if is better to use perturbation (to reduce the chance of producing a non-manifold solid).