Improved robutsness
Reclassification stage now uses two sets of reclassification values. prev_position is the source for the analysis of the reclassification stage, recl_position is the result.

During a reclassification stage, recl_position values are used only to set the intersection flag of the sector to false.

From one reclassification stage to the next, prev_position values are updated with recl_position values.

At last, recl_position values are used to insert the null edges.