Method to get the real bbox extension of a solid.
Solid bbox was based on an extended bbox to allow false positives in intersection tests. Computing a real bbox was unfeasible from this data. Now, data in bbox is splitted in two parts, the real and the extended coordinates of the bbox.

The extended part is used for all geometric test. The real part is queried by the new method.